+++ spacecookie.nix
@@ -1,8 +1,7 @@
 { mkDerivation, aeson, async, attoparsec, base, bytestring
-, containers, directory, download-curl, fast-logger
-, filepath-bytestring, lib, mtl, process, socket, systemd, tasty
-, tasty-expected-failure, tasty-hunit, text, transformers, unix
-, utf8-string
+, containers, directory, download-curl, fast-logger, file-io
+, filepath, lib, mtl, os-string, process, socket, systemd, tasty
+, tasty-expected-failure, tasty-hunit, text, unix, utf8-string
 }:
 mkDerivation {
   pname = "spacecookie";
@@ -11,21 +10,19 @@ mkDerivation {
   isLibrary = true;
   isExecutable = true;
   libraryHaskellDepends = [
-    async attoparsec base bytestring containers directory
-    filepath-bytestring mtl socket text transformers unix utf8-string
+    async attoparsec base bytestring containers filepath mtl os-string
+    socket text utf8-string
   ];
   executableHaskellDepends = [
     aeson attoparsec base bytestring containers directory fast-logger
-    filepath-bytestring mtl socket systemd text transformers unix
-    utf8-string
+    file-io filepath os-string socket systemd text unix utf8-string
   ];
   testHaskellDepends = [
     attoparsec base bytestring containers directory download-curl
-    filepath-bytestring process tasty tasty-expected-failure
-    tasty-hunit utf8-string
+    filepath os-string process tasty tasty-expected-failure tasty-hunit
   ];
   homepage = "https://github.com/sternenseemann/spacecookie";
   description = "Gopher server library and daemon";
-  license = lib.licenses.gpl3Only;
+  license = lib.licensesSpdx."GPL-3.0-only";
   mainProgram = "spacecookie";
 }